Changed the title to include SASS
SASS is a very useful tool and makes CSS more readable. SASS allows nesting syntax which groups selectors together in an intuitive way, and SASS allows variables which are useful for having a single source of truth.

If anything should be taught about SASS, it should be nesting syntax and variables. If anything should be taught about Tailwind, it should be utility classes.

Doing a job search, there are very few jobs that require Tailwind, but a large number that require SASS. It makes sense to spend more time teaching SASS because more jobs require it.